Fix Yaml issues

+++ include/gnss/data_distributor.h
@@ -11,22 +11,19 @@
 #include <gnss/api.h>
 #include <gnuradio/sync_block.h>
 
-namespace gr
-{
-  namespace gnss
-  {
+namespace gr {
+namespace gnss {
 
-    /*!
+/*!
  * \brief <+description of block+>
  * \ingroup gnss
  *
  */
-    class GNSS_API data_distributor : virtual public gr::sync_block
-    {
-    public:
-      typedef std::shared_ptr<data_distributor> sptr;
+class GNSS_API data_distributor : virtual public gr::sync_block {
+public:
+  typedef std::shared_ptr<data_distributor> sptr;
 
-      /*!
+  /*!
    * \brief Return a shared_ptr to a new instance of gnss::data_distributor.
    *
    * To avoid accidental use of raw pointers, gnss::data_distributor's
@@ -34,10 +31,10 @@ namespace gr
    * class. gnss::data_distributor::make is the public interface for
    * creating new instances.
    */
-      static sptr make(unsigned int numSamples);
-    };
+  static sptr make(float numSamples);
+};
 
-  } // namespace gnss
+} // namespace gnss
 } // namespace gr
 
 #endif /* INCLUDED_GNSS_DATA_DISTRIBUTOR_H */
